+++ HISTORY.rst
@@ -2,3 +2,8 @@
 ------------------
 
 * First release on PyPI.
+
+0.2.0 (2022-05-02)
+----------------------
+* Add initial support for CPCs, LPARs, and logical CPUs
+* Change minimum python level to 3.7 so I can use dataclasses.  3.6 is EOL anyway.